requirements.txt
# langchain
# python-dotenv
# ipykernel
# langchain_community
# PyPDF
# bs4
# arxiv
# PyMuPDF
# wikipedia
# langchain-text-splitters            
# lxml
# langchain-community
# sentence_transformers
# langchain_huggingface
# faiss-cpu
# chromadb
# langchain_chroma
# streamlit
# langchain_groq
# langchain_core
# fastapi
# uvicorn
# langserve[all]
# langchain_community
# ipywidgets
# tf-keras

langchain
python-dotenv
ipykernel
langchain_community
PyPDF
bs4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters            
lxml
langchain-community
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve[all]
langchain_community
ipywidgets
tf-keras
langchain-nvidia-ai-endpoints
langchain-google-genai
qdrant_client
langchain_qdrant